+++ base/src/main/java/io/github/sds100/keymapper/base/actions/ConfigActionsViewModel.kt
@@ -21,6 +21,7 @@ import io.github.sds100.keymapper.base.utils.ui.DialogProvider
 import io.github.sds100.keymapper.base.utils.ui.ResourceProvider
 import io.github.sds100.keymapper.base.utils.ui.ViewModelHelper
 import io.github.sds100.keymapper.base.utils.ui.compose.ComposeIconInfo
+import io.github.sds100.keymapper.base.vibration.VibrateConfigDelegate
 import io.github.sds100.keymapper.common.utils.AccessibilityServiceError
 import io.github.sds100.keymapper.common.utils.KMError
 import io.github.sds100.keymapper.common.utils.State
@@ -53,6 +54,7 @@ class ConfigActionsViewModel @Inject constructor(
     setupAccessibilityServiceDelegate: SetupAccessibilityServiceDelegate,
     fixKeyEventActionDelegate: FixKeyEventActionDelegate,
     private val onboardingTipDelegate: OnboardingTipDelegate,
+    private val vibrateConfigDelegate: VibrateConfigDelegate,
     resourceProvider: ResourceProvider,
     navigationProvider: NavigationProvider,
     dialogProvider: DialogProvider,
@@ -67,7 +69,7 @@ class ConfigActionsViewModel @Inject constructor(
     OnboardingTipDelegate by onboardingTipDelegate {
 
     val createActionDelegate =
-        CreateActionDelegate(viewModelScope, createAction, this, this, this)
+        CreateActionDelegate(viewModelScope, createAction, this, this, this, vibrateConfigDelegate)
     private val uiHelper = ActionUiHelper(displayAction, resourceProvider)
 
     private val _state = MutableStateFlow<State<ConfigActionsState>>(State.Loading)