+++ base/src/main/java/io/github/sds100/keymapper/base/detection/DetectKeyMapsUseCase.kt
@@ -19,6 +19,8 @@ import io.github.sds100.keymapper.base.system.accessibility.IAccessibilityServic
 import io.github.sds100.keymapper.base.system.navigation.OpenMenuHelper
 import io.github.sds100.keymapper.base.trigger.FingerprintTriggerKey
 import io.github.sds100.keymapper.base.utils.ui.ResourceProvider
+import io.github.sds100.keymapper.base.vibration.VibrateEffect
+import io.github.sds100.keymapper.base.vibration.vibrate
 import io.github.sds100.keymapper.common.utils.State
 import io.github.sds100.keymapper.common.utils.dataOrNull
 import io.github.sds100.keymapper.data.Keys
@@ -181,6 +183,10 @@ class DetectKeyMapsUseCaseImpl @AssistedInject constructor(
         vibrator.vibrate(duration)
     }
 
+    override fun vibrate(effect: VibrateEffect) {
+        vibrator.vibrate(effect)
+    }
+
     override fun imitateKeyEvent(
         keyCode: Int,
         metaState: Int,
@@ -267,6 +273,7 @@ interface DetectKeyMapsUseCase {
 
     fun showTriggeredToast()
     fun vibrate(duration: Long)
+    fun vibrate(effect: VibrateEffect)
 
     val currentTime: Long