+++ base/src/main/java/io/github/sds100/keymapper/base/vibration/VibrateEffect.kt
@@ -1,0 +1,20 @@
+package io.github.sds100.keymapper.base.vibration
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+sealed class VibrateEffect {
+    @Serializable
+    data class CustomDuration(val durationMs: Long) : VibrateEffect()
+
+    @Serializable
+    data class Predefined(val predefinedType: PredefinedType) : VibrateEffect()
+
+    @Serializable
+    enum class PredefinedType {
+        CLICK,
+        DOUBLE_CLICK,
+        HEAVY_CLICK,
+        TICK,
+    }
+}