+++ base/src/test/java/io/github/sds100/keymapper/base/actions/PerformActionsUseCaseTest.kt
@@ -91,6 +91,7 @@ class PerformActionsUseCaseTest {
             coroutineScope = testCoroutineScope,
             notificationAdapter = mock(),
             settingsAdapter = mock(),
+            vibratorAdapter = mock(),
         )
     }