+++ system/src/main/java/io/github/sds100/keymapper/system/vibrator/PredefinedVibrationEffect.kt
@@ -1,0 +1,8 @@
+package io.github.sds100.keymapper.system.vibrator
+
+enum class PredefinedVibrationEffect {
+    CLICK,
+    DOUBLE_CLICK,
+    HEAVY_CLICK,
+    TICK,
+}